# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/menu/": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call GetMenu, return ok by json.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call UpdateMenu, return ok by json.",
                "parameters": [
                    {
                        "description": "menuInput",
                        "name": "menu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.MenuInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            },
            "post": {
                "description": "DB에 menu 추가 []",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call InsertMenu, return ok by json.",
                "parameters": [
                    {
                        "description": "menuInput",
                        "name": "menu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.MenuInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            },
            "delete": {
                "description": "DB에 menu 추가",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call DeleteMenu, return ok by json.",
                "parameters": [
                    {
                        "description": "menuName",
                        "name": "menuName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            }
        },
        "/menu/admin": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call RecommendMenu, return ok by json.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "recommend",
                        "name": "recommend",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            }
        },
        "/menu/grade": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call SortMenu, return ok by json.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            }
        },
        "/order/": {
            "get": {
                "description": "유저별 주문 내역 조회",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call GetOrderByUser, return ok by json.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pnum",
                        "name": "pnum",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            },
            "put": {
                "description": "유저 추가 주문",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call AddOrderMenu, return ok by json.",
                "parameters": [
                    {
                        "description": "InputOrder",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.InputOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            },
            "post": {
                "description": "DB에 주문 추가",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call InsertOrder, return ok by json.",
                "parameters": [
                    {
                        "description": "InputOrder",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.InputOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            }
        },
        "/order/state": {
            "put": {
                "description": "주문 상태를 변경",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call UpdateOrderState, return ok by json.",
                "parameters": [
                    {
                        "description": "InputOrderState",
                        "name": "orderstate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.InputOrderState"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            }
        },
        "/review/reviewer": {
            "get": {
                "description": "해당 메뉴의 평균 평점과 리뷰를 조회",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call GetReviewWithMenu, return ok by json.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "menuName",
                        "name": "menuname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            },
            "put": {
                "description": "메뉴의 평점 및 리뷰를 남김",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "call UpdateMenuGrade, return ok by json.",
                "parameters": [
                    {
                        "description": "InputGrade",
                        "name": "review",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.InputGrade"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Controller"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.Controller": {
            "type": "object"
        },
        "controller.InputGrade": {
            "type": "object",
            "properties": {
                "grade": {
                    "type": "integer"
                },
                "menuName": {
                    "type": "string"
                },
                "orderNumber": {
                    "type": "integer"
                },
                "orderTime": {
                    "type": "string"
                },
                "review": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
        "controller.InputOrder": {
            "type": "object",
            "properties": {
                "menuInfo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.MenuInfo"
                    }
                },
                "user": {
                    "$ref": "#/definitions/controller.InputUser"
                }
            }
        },
        "controller.InputOrderState": {
            "type": "object",
            "properties": {
                "number": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "controller.InputUser": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "pnum": {
                    "type": "string"
                }
            }
        },
        "controller.MenuInfo": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "controller.MenuInput": {
            "type": "object",
            "properties": {
                "isOrder": {
                    "type": "boolean"
                },
                "isVisible": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "recommend": {
                    "type": "boolean"
                },
                "spicy": {
                    "type": "integer"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "pnum": {
                    "type": "string"
                }
            }
        }
    }
}